+++ frontend/src/componments/Button.js
@@ -1,15 +1,18 @@
 import React from 'react';
+import {motion} from 'framer-motion';
 import './Button.css'
 
 function Button({ text, disabled=false, onClick=() => {}, style={}}) {
     return (
-        <button
+        <motion.button
+            whileHover={{ scale: 1.1 }}
+            whileTap={{ scale: 0.9 }}
             className={ disabled ? "button-disabled" : "button" }
             onClick={onClick}
             style={style}
         >
             {text}
-        </button>
+        </motion.button>
     );
 }